+++ Runtime/Internal/ChoiceState.cs
@@ -19,7 +19,7 @@ namespace GameLovers.Statechart.Internal
 		/// <inheritdoc />
 		public override void Enter()
 		{
-			for(int i = 0; i < _onEnter.Count; i++)
+			for(var i = 0; i < _onEnter.Count; i++)
 			{
 				_onEnter[i]?.Invoke();
 			}
@@ -28,7 +28,7 @@ namespace GameLovers.Statechart.Internal
 		/// <inheritdoc />
 		public override void Exit()
 		{
-			for(int i = 0; i < _onExit.Count; i++)
+			for(var i = 0; i < _onExit.Count; i++)
 			{
 				_onExit[i]?.Invoke();
 			}
@@ -40,7 +40,7 @@ namespace GameLovers.Statechart.Internal
 #if UNITY_EDITOR || DEBUG
 			int noTransitionConditionCount = 0;
 
-			for(int i = 0; i < _transitions.Count; i++)
+			for(var i = 0; i < _transitions.Count; i++)
 			{
 				if (!_transitions[i].HasCondition)
 				{
@@ -103,7 +103,7 @@ namespace GameLovers.Statechart.Internal
 		{
 			ITransitionInternal noTransitionCondition = null;
 
-			for(int i = 0; i < _transitions.Count; i++)
+			for(var i = 0; i < _transitions.Count; i++)
 			{
 				if (!_transitions[i].HasCondition)
 				{