+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-08-27
+
+- Added the possibility to always execute the *FinalState* of a *NestState* and * *SplitState*
+
+**Changed**
+- Splitted the Unit tests in different files to help organization and readability. Now each state has it's own file with only it's set of tests
+
+**Fixed**
+- Now *NestState* and * *SplitState* properly execute their inner *OnExit* calls when they are setup in a chain of nested states
+
 ## [0.1.3] - 2020-01-06
 
 - Removed package dependency